Add in unittests and refactor directory structure
