+++ common/document_parser/lib/responsibility_parse/__init__.py
@@ -1,1 +1,1 @@
-from .parse_responsibilties import ResponsibilityParser
+from .responsibility_parser import ResponsibilityParser
